Implement msp parsing, update existing tests and spread functions across additional files
