+++ tests/test_data/massbank_msp.txt
@@ -1,0 +1,87 @@
+ACCESSION: UO000002
+RECORD_TITLE: 2,3-di-O-Phytanyl-sn-glycerol-1-phosphoserine; EI-B; MS
+DATE: 2016.01.19 (Created 2009.05.29, modified 2011.05.06)
+AUTHORS: Hiroyuki Morii, Department of Chemistry, University of Occupational and Environmental Health
+LICENSE: CC BY-SA
+PUBLICATION: Morii, H., Nishihara, M., Ohga, M., and Koga, Y. 1986. A diphytanyl ether analog of phosphatidylserine from a methanogenic bacterium, Methanobrevibacter arboriphilus. J Lipid Res. 27: 724-730.
+COMMENT: Ammonium salt of the compound was analyzed
+COMMENT: [Analytical] Ionizing Curr 300 uA, Chamber Temp 250 C, Accel Volt 3KV, Ion Multi 1.0 KV,
+CH$NAME: 2,3-di-O-Phytanyl-sn-glycerol-1-phosphoserine
+CH$NAME: archaetidylserine
+CH$COMPOUND_CLASS: Glycerophospholipids; Glycerophosphoserines; Dialkylglycerophosphoserines
+CH$FORMULA: C46H94NO8P
+CH$EXACT_MASS: 819.67171
+CH$SMILES: C(CCC(C)C)C(C)CCCC(CCCC(CCOCC(OCCC(CCCC(C)CCCC(C)CCCC(C)C)C)COP(O)(=O)OCC(C(O)=O)N)C)C
+CH$IUPAC: InChI=1S/C46H94NO8P/c1-36(2)17-11-19-38(5)21-13-23-40(7)25-15-27-42(9)29-31-52-33-44(34-54-56(50,51)55-35-45(47)46(48)49)53-32-30-43(10)28-16-26-41(8)24-14-22-39(6)20-12-18-37(3)4/h36-45H,11-35,47H2,1-10H3,(H,48,49)(H,50,51)/t38-,39-,40-,41-,42-,43-,44-,45-/m1/s1
+CH$LINK: CAS 105662-26-8
+CH$LINK: LIPIDBANK EEL3026
+AC$INSTRUMENT: JMS DX-300/JMS-3500 data system, Japan Electron Optics Laboratory, Japan
+AC$INSTRUMENT_TYPE: EI-B
+AC$MASS_SPECTROMETRY: MS_TYPE MS
+AC$MASS_SPECTROMETRY: ION_MODE POSITIVE
+AC$MASS_SPECTROMETRY: IONIZATION_POTENTIAL 30 eV
+AC$MASS_SPECTROMETRY: SCANNING 5 Sec
+AC$MASS_SPECTROMETRY: SOURCE_TEMPERATURE 320 C
+MS$FOCUSED_ION: ION_TYPE [M]+*
+PK$SPLASH: splash10-05gi-9611001000-5e7663b41cf47681770e
+PK$NUM_PEAK: 58
+PK$PEAK: m/z int. rel.int.
+  36.0 48.935 69
+  43.0 155.642 221
+  55.0 174.853 248
+  56.0 241.397 343
+  57.0 685.069 973
+  69.0 429.724 610
+  70.0 442.816 629
+  71.0 703.562 999
+  74.0 153.368 218
+  81.0 183.718 261
+  82.0 116.090 165
+  83.0 432.501 614
+  84.0 189.394 269
+  85.0 524.513 745
+  95.0 102.128 145
+  96.0 179.720 255
+  97.0 449.439 638
+  98.0 118.287 168
+  99.0 376.928 535
+  111.0 429.907 610
+  112.0 168.627 239
+  113.0 308.186 438
+  123.0 298.191 423
+  124.0 233.188 331
+  125.0 504.097 716
+  126.0 362.310 514
+  127.0 300.450 427
+  139.0 123.338 175
+  140.0 233.340 331
+  141.0 235.767 335
+  153.0 107.469 153
+  155.0 163.241 232
+  169.0 130.189 185
+  182.0 94.773 135
+  183.0 160.234 228
+  196.0 132.555 188
+  197.0 163.622 232
+  211.0 75.654 107
+  278.0 326.649 464
+  279.0 220.386 313
+  280.0 227.649 323
+  281.0 143.572 204
+  296.0 158.358 225
+  297.0 60.502 86
+  309.0 35.370 50
+  325.0 279.819 397
+  326.0 71.000 101
+  340.0 133.760 190
+  341.0 60.486 86
+  343.0 243.579 346
+  344.0 60.028 85
+  354.0 51.468 73
+  373.0 132.555 188
+  374.0 42.069 60
+  383.0 63.767 91
+  634.0 450.446 640
+  635.0 212.497 302
+  636.0 49.622 70
+//
