+++ tests/test_isomorphism_database.py
@@ -21,6 +21,7 @@
 
 
 import os
+import sys
 import unittest
 import shutil
 import tempfile
