+++ tests/test_suite_auxiliary.py
@@ -27,6 +27,7 @@ import unittest
 from pathlib import Path
 
 from . import test_auxiliary
+from . import test_parse
 
 sys.path.insert(0, str(Path(__file__).parent.parent.resolve()))
 
@@ -35,6 +36,7 @@ if __name__ == '__main__':
     suite = unittest.TestSuite()
 
     suite.addTest(unittest.findTestCases(test_auxiliary))
+    suite.addTest(unittest.findTestCases(test_parse))
 
     report = os.path.join(os.path.abspath(os.path.join(__file__, os.pardir)), 'results', 'results_test_suite_auxiliary')
     runTestSuite(suite, report, title='Process Test Suite Report', verbosity=2)
